Fix a rare privacy bug in DistinctPerKey in Privacy on Beam.
The bug occurred when there are outlier users in the input that
contribute to many partitions and/or to many values AND the
values contributed are the same as values from other users (the second
part is critical, if the contributed values only come from a single user
then the bug does not occur). Then, the output might not have be DP due
to incorrect contribution bounding. See the comments in the newly added
tests for concrete examples of when/how the bug used to occur.

GitOrigin-RevId: 27bd7a0f5ccb16fe5bd0acda5d141713262638ab
Change-Id: I68c3765e95a6b97ed537d1c188736b28b776417e