Fix sensitivity calculation in C++
C++ DP library:
- Fix privacy-impacting bug in sensitivity calculation for integer mean.
We found a rounding issue in the integer version for BoundedMean when
abs(upper-lower) is odd.  In this case, the sensitivity was rounded down to the next integer. In case that abs(upper-lower) = 1, the builder would have returned an error. For other bounds, the sensitivity calculation is slightly off and BoundedMean produces an output with slightly higher epsilon and delta. This does not apply to floating point types.
- Add default expectations to the mock numerical mechanism.
The defaults will redirect calls to the single-argument versions of the Add*Noise methods, allowing users of the mock to set expectations on the single-argument versions. These will be removed when we remove the multi-argument shims.
- Fix message handling of optional value when value is unset. Add tests for unset and NaN values for lt, lte, gt, and gte methods.
- Add confidence interval for fixed-bounds mean

Python Accounting library:
- Fix libraries versions to avoid runtime errors.

Privacy-on-Beam
- Fix tolerance computation for mean, improve documentation.

PiperOrigin-RevId: 445993700
Change-Id: Iccc2938730fdd5de79114e09c178cec5f6692771
GitOrigin-RevId: ec51b89721f1e555c2b6370b794b329f8f47c34a