C++ change return types of NumericalMechanism::AddNoise
We are changing the return type of NumericalMechanism::AddNoise from
T to double or int64_t.  This reduces confusion when T is an unsigned
type.  This is *not* a breaking change, since there is an implicit
conversion and when using T, which was in the library, and is now in
control of the user.

Additionally, this commit contains minor clean-ups and an absl
dependency update.

PiperOrigin-RevId: 634249123
Change-Id: I8423326fb6056e944fbc4d9ceae2c8918edcfedc
GitOrigin-RevId: b9193e7d7621855b41d1b430ec01d4534b73f407